COPS-805: Replace the `Process` abstraction with an async connection.
This does a few things. First; it gets rid of a hacky abstraction that's
a bit unintuitive for one that's much more common. Second; it's the
first of a few tasks that allow us to step away from the
`threading.Thread` approach to asynchronous handling in favor of the
mature and well-loved tornado.ioloop module. Finally; This sets the
scene  to further expose an asynchronous interface to the end user so there
won't be annoying waits when `Update` objects are first incorporated
into the `LruCache`.